+++ R/allTabsBP.4.R
@@ -1,11 +1,11 @@
-#' An example of "tableList" object resulting from a call to 'buildEnrichTable'
+#' An example of an object of class "tableList" resulting from a call to 'buildEnrichTable'
 #'
 #' The result of generating all contingency tables of mutual enrichment, in a pairwise fashion,
 #' between the gene lists in data \code{\link{allOncoGeneLists}}. The information in these data
 #' was summarized as 2x2 contingency tables of GO items enrichment, at level 4 of the BP
-#' ontology. These results are based on gene lists which are non automatically updated, take
-#' them just as an illustrative example because the gene lists, and the GO, may change along
-#' time.
+#' ontology. These results are based on gene lists which are not automatically updated, take
+#' them just as an illustrative example because the gene lists, the GO and Bioconductor, may change
+#' along time. The present version of these data was generated under Bioconductor version 3.17.
 #'
 #' @format An object of class "tableList" inheriting from class "list". It is a list of class
 #' "table" objects.