+++ R/boot.cancerEquivSorensen.R
@@ -1,9 +1,9 @@
-#' An example of "AllEquivSDhtest" object resulting from a call to 'allEquivTestSorensen'
+#' An example of object of class "AllEquivSDhtest" resulting from a call to 'allEquivTestSorensen'
 #'
 #' The bootstrap Sorensen-Dice test performed on the cancer gene lists in data \code{\link{allOncoGeneLists}}
-#' which is automatically charged with this package. The test is iterated for all GO ontologies and for GO
-#' levels 3 to 10. These results are not automatically updated for changes in these gene lists and Bioconductor
-#' and Go updates, take them just as an illustrative example.
+#' which is may be charged from this package. The test is iterated for all GO ontologies and for GO
+#' levels 3 to 10. These results are not automatically updated for changes in these gene lists, take them just
+#' as an illustrative example. The present version was obtained under Bioconductor 3.17.
 #'
 #' For each ontology and GO level, the result contains the result of all pairwise tests of equivalence between
 #' the cancer gene lists.