+++ R/tab_atlas.sanger_BP3.R
@@ -6,7 +6,7 @@
 #' Enriched in both lists, enriched in sanger but not in atlas, non-enriched in sanger
 #' but enriched in atlas and non-enriched in both lists.
 #' Take it just as an illustrative example, non up-to-date for changes in the gene lists
-#' or changes in the GO.
+#' or changes in the GO. The present version was obtained under Bioconductor 3.17.
 #'
 #' @format An object of class "table" representing a 2x2 contingency table.
 #' @source \url{http://www.bushmanlab.org/links/genelists}